Market Item View screen logic. Inventory landing page logic. Restructure architecture
IMplement the logic for the Market-Item-View screen along with navigation.
Create the Inventory first page to display options to View or Add Inventory.
Refactor the logic for displaying an image from the server to a standalone Composable. Thus cleaning it up and allowing it to be reused in other places.
Restructure the project to better follow Clean Architecture. Including adding a "domain" layer, Dto -> Domain Model mapping
Add Chat bottom nav bar option. Not implemented yet.
Swapped out Market bottom nav bar icon for one that more clearly shows when it is selected